SEMESTER 1/Algoritma dan Struktur Data 1/Code/array/multidimensi/ular-tangga/requirements.txt
pillow
tkinker